service

+++ lualib-src/mtask_lua_socket.c
@@ -15,7 +15,7 @@
 #include "mtask_socket.h"
 
 #define BACKLOG 32
-// 2 ** 12 == 4096
+// 2 ^ 12 == 4096
 #define LARGE_PAGE_NODE 12
 #define BUFFER_LIMIT (256 * 1024)
 
@@ -35,7 +35,7 @@ struct socket_buffer {
 static int
 lfreepool(lua_State *L) {
 	struct buffer_node * pool = lua_touserdata(L, 1);
-	int sz = lua_rawlen(L,1) / sizeof(*pool);
+	int sz = (int)(lua_rawlen(L,1) / sizeof(*pool));
 	int i;
 	for (i=0;i<sz;i++) {
 		struct buffer_node *node = &pool[i];
@@ -105,12 +105,12 @@ lpushbuffer(lua_State *L) {
 	}
 	int pool_index = 2;
 	luaL_checktype(L,pool_index,LUA_TTABLE);
-	int sz = luaL_checkinteger(L,4);
+	int sz = (int)luaL_checkinteger(L,4);
 	lua_rawgeti(L,pool_index,1);
 	struct buffer_node * free_node = lua_touserdata(L,-1);	// sb poolt msg size free_node
 	lua_pop(L,1);
 	if (free_node == NULL) {
-		int tsz = lua_rawlen(L,pool_index);
+		int tsz = (int)lua_rawlen(L,pool_index);
 		if (tsz == 0)
 			tsz++;
 		int size = 8;
@@ -235,7 +235,7 @@ lpopbuffer(lua_State *L) {
 		return luaL_error(L, "Need buffer object at param 1");
 	}
 	luaL_checktype(L,2,LUA_TTABLE);
-	int sz = luaL_checkinteger(L,3);
+	int sz = (int)luaL_checkinteger(L,3);
 	if (sb->size < sz || sz == 0) {
 		lua_pushnil(L);
 	} else {
@@ -333,11 +333,11 @@ lreadline(lua_State *L) {
 	int from = sb->offset;
 	int bytes = current->sz - from;
 	for (i=0;i<=sb->size - (int)seplen;i++) {
-		if (check_sep(current, from, sep, seplen)) {
+		if (check_sep(current, from, sep, (int)seplen)) {
 			if (check) {
 				lua_pushboolean(L,true);
 			} else {
-				pop_lstring(L, sb, i+seplen, seplen);
+				pop_lstring(L, sb, (int)(i+seplen), (int)seplen);
 				sb->size -= i+seplen;
 			}
 			return 1;
@@ -377,7 +377,7 @@ lstr2p(lua_State *L) {
 static int
 lunpack(lua_State *L) {
 	struct mtask_socket_message *message = lua_touserdata(L,1);
-	int size = luaL_checkinteger(L,2);
+	int size = (int)luaL_checkinteger(L,2);
 
 	lua_pushinteger(L, message->type);
 	lua_pushinteger(L, message->id);
@@ -417,7 +417,7 @@ address_port(lua_State *L, char *tmp, const char * addr, int port_index, int *po
 			if (sep == NULL) {
 				luaL_error(L, "Invalid address %s.",addr);
 			}
-			*port = strtoul(sep+1,NULL,10);
+			*port = (int)strtoul(sep+1,NULL,10);
 		} else {
 			// is ipv4
 			const char * sep = strchr(addr,':');
@@ -427,11 +427,11 @@ address_port(lua_State *L, char *tmp, const char * addr, int port_index, int *po
 			memcpy(tmp, addr, sep-addr);
 			tmp[sep-addr] = '\0';
 			host = tmp;
-			*port = strtoul(sep+1,NULL,10);
+			*port = (int)strtoul(sep+1,NULL,10);
 		}
 	} else {
 		host = addr;
-		*port = luaL_optinteger(L,port_index, 0);
+		*port = (int)luaL_optinteger(L,port_index, 0);
 	}
 	return host;
 }
@@ -455,7 +455,7 @@ lconnect(lua_State *L) {
 
 static int
 lclose(lua_State *L) {
-	int id = luaL_checkinteger(L,1);
+	int id = (int)luaL_checkinteger(L,1);
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
 	mtask_socket_close(ctx, id);
 	return 0;
@@ -464,8 +464,8 @@ lclose(lua_State *L) {
 static int
 llisten(lua_State *L) {
 	const char * host = luaL_checkstring(L,1);
-	int port = luaL_checkinteger(L,2);
-	int backlog = luaL_optinteger(L,3,BACKLOG);
+	int port = (int)luaL_checkinteger(L,2);
+	int backlog = (int)luaL_optinteger(L,3,BACKLOG);
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
 	int id = mtask_socket_listen(ctx, host,port,backlog);
 	if (id < 0) {
@@ -481,7 +481,7 @@ get_buffer(lua_State *L, int index, int *sz) {
 	void *buffer;
 	if (lua_isuserdata(L,index)) {
 		buffer = lua_touserdata(L,index);
-		*sz = luaL_checkinteger(L,index+1);
+		*sz = (int)luaL_checkinteger(L,index+1);
 	} else {
 		size_t len = 0;
 		const char * str =  luaL_checklstring(L, index, &len);
@@ -495,7 +495,7 @@ get_buffer(lua_State *L, int index, int *sz) {
 static int
 lsend(lua_State *L) {
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
-	int id = luaL_checkinteger(L, 1);
+	int id = (int)luaL_checkinteger(L, 1);
 	int sz = 0;
 	void *buffer = get_buffer(L, 2, &sz);
 	int err = mtask_socket_send(ctx, id, buffer, sz);
@@ -506,7 +506,7 @@ lsend(lua_State *L) {
 static int
 lsendlow(lua_State *L) {
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
-	int id = luaL_checkinteger(L, 1);
+	int id = (int)luaL_checkinteger(L, 1);
 	int sz = 0;
 	void *buffer = get_buffer(L, 2, &sz);
 	mtask_socket_send_lowpriority(ctx, id, buffer, sz);
@@ -516,7 +516,7 @@ lsendlow(lua_State *L) {
 static int
 lbind(lua_State *L) {
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
-	int fd = luaL_checkinteger(L, 1);
+	int fd = (int)luaL_checkinteger(L, 1);
 	int id = mtask_socket_bind(ctx,fd);
 	lua_pushinteger(L,id);
 	return 1;
@@ -525,7 +525,7 @@ lbind(lua_State *L) {
 static int
 lstart(lua_State *L) {
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
-	int id = luaL_checkinteger(L, 1);
+	int id = (int)luaL_checkinteger(L, 1);
 	mtask_socket_start(ctx,id);
 	return 0;
 }
@@ -533,7 +533,7 @@ lstart(lua_State *L) {
 static int
 lnodelay(lua_State *L) {
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
-	int id = luaL_checkinteger(L, 1);
+	int id = (int)luaL_checkinteger(L, 1);
 	mtask_socket_nodelay(ctx,id);
 	return 0;
 }
@@ -561,7 +561,7 @@ ludp(lua_State *L) {
 static int
 ludp_connect(lua_State *L) {
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
-	int id = luaL_checkinteger(L, 1);
+	int id = (int)luaL_checkinteger(L, 1);
 	size_t sz = 0;
 	const char * addr = luaL_checklstring(L,2,&sz);
 	char tmp[sz];
@@ -581,7 +581,7 @@ ludp_connect(lua_State *L) {
 static int
 ludp_send(lua_State *L) {
 	struct mtask_context * ctx = lua_touserdata(L, lua_upvalueindex(1));
-	int id = luaL_checkinteger(L, 1);
+	int id = (int)luaL_checkinteger(L, 1);
 	const char * address = luaL_checkstring(L, 2);
 	int sz = 0;
 	void *buffer = get_buffer(L, 3, &sz);
